Adding Irrationals to Expression works. 2 * pi crashed. Now yields 2*pi (pi + 6) + pi crashed. Now yields 2*pi + 6. Adding negative rationals or irrationals also works. However, if the result is 0 then 0 will either stay there or the program will crash.
Had to modify the coefficient stuff a bit in IrrationalNumber and then
update Add to deal with the adding.
Also, updated Multiply to handle IrrationalNumbers and RationalNumbers
only. (Adding Expression support is next)
IrrationalNumber still needs Value* coefficient support. I know it's
going to be a bitch but it's the only way.
We can probably keep the coefficient code with integers. Just add the
other stuff around that.